🎨 Collect together the package-related utils
- read directory actually reads a directory of packages
- parse package json is very tighly related to this
