+++ core/test/unit/server_utils_spec.js
@@ -1,14 +1,8 @@
 var should          = require('should'),
     sinon           = require('sinon'),
     nock            = require('nock'),
-    parsePackageJson = require('../../server/utils/parse-package-json'),
-    readDirectory   = require('../../server/utils/read-directory'),
-    readThemes      = require('../../server/utils/read-themes'),
     gravatar        = require('../../server/utils/gravatar'),
-    tmp             = require('tmp'),
-    utils           = require('../../server/utils'),
-    join            = require('path').join,
-    fs              = require('fs');
+    utils           = require('../../server/utils');
 
 // To stop jshint complaining
 should.equal(true, true);
@@ -97,272 +91,6 @@ describe('Server Utilities', function () {
                 options = {importing: true};
             result = safeString('-slug-&with-✓-invalid-characters-に\'', options);
             result.should.equal('-slug--with--invalid-characters-ni');
-        });
-    });
-
-    describe('parse-package-json', function () {
-        it('should parse valid package.json', function (done) {
-            var pkgJson, tmpFile;
-
-            tmpFile = tmp.fileSync();
-            pkgJson = JSON.stringify({
-                name: 'test',
-                version: '0.0.0'
-            });
-
-            fs.writeSync(tmpFile.fd, pkgJson);
-
-            parsePackageJson(tmpFile.name)
-                .then(function (pkg) {
-                    pkg.should.eql({
-                        name: 'test',
-                        version: '0.0.0'
-                    });
-
-                    done();
-                })
-                .catch(done)
-                .finally(tmpFile.removeCallback);
-        });
-
-        it('should fail when name is missing', function (done) {
-            var pkgJson, tmpFile;
-
-            tmpFile = tmp.fileSync();
-            pkgJson = JSON.stringify({
-                version: '0.0.0'
-            });
-
-            fs.writeSync(tmpFile.fd, pkgJson);
-
-            parsePackageJson(tmpFile.name)
-                .then(function () {
-                    done(new Error('parsePackageJson succeeded, but should\'ve failed'));
-                })
-                .catch(function (err) {
-                    err.message.should.equal('"name" or "version" is missing from theme package.json file.');
-                    err.context.should.equal(tmpFile.name);
-                    err.help.should.equal('This will be required in future. Please see http://docs.ghost.org/themes/');
-
-                    done();
-                })
-                .catch(done)
-                .finally(tmpFile.removeCallback);
-        });
-
-        it('should fail when version is missing', function (done) {
-            var pkgJson, tmpFile;
-
-            tmpFile = tmp.fileSync();
-            pkgJson = JSON.stringify({
-                name: 'test'
-            });
-
-            fs.writeSync(tmpFile.fd, pkgJson);
-
-            parsePackageJson(tmpFile.name)
-                .then(function () {
-                    done(new Error('parsePackageJson succeeded, but should\'ve failed'));
-                })
-                .catch(function (err) {
-                    err.message.should.equal('"name" or "version" is missing from theme package.json file.');
-                    err.context.should.equal(tmpFile.name);
-                    err.help.should.equal('This will be required in future. Please see http://docs.ghost.org/themes/');
-
-                    done();
-                })
-                .catch(done)
-                .finally(tmpFile.removeCallback);
-        });
-
-        it('should fail when JSON is invalid', function (done) {
-            var pkgJson, tmpFile;
-
-            tmpFile = tmp.fileSync();
-            pkgJson = '{name:"test"}';
-
-            fs.writeSync(tmpFile.fd, pkgJson);
-
-            parsePackageJson(tmpFile.name)
-                .then(function () {
-                    done(new Error('parsePackageJson succeeded, but should\'ve failed'));
-                })
-                .catch(function (err) {
-                    err.message.should.equal('Theme package.json file is malformed');
-                    err.context.should.equal(tmpFile.name);
-                    err.help.should.equal('This will be required in future. Please see http://docs.ghost.org/themes/');
-
-                    done();
-                })
-                .catch(done)
-                .finally(tmpFile.removeCallback);
-        });
-
-        it('should fail when file is missing', function (done) {
-            var tmpFile = tmp.fileSync();
-
-            tmpFile.removeCallback();
-            parsePackageJson(tmpFile.name)
-                .then(function () {
-                    done(new Error('parsePackageJson succeeded, but should\'ve failed'));
-                })
-                .catch(function (err) {
-                    err.message.should.equal('Could not read package.json file');
-                    err.context.should.equal(tmpFile.name);
-
-                    done();
-                })
-                .catch(done);
-        });
-    });
-
-    describe('read-directory', function () {
-        it('should read directory recursively', function (done) {
-            var themePath = tmp.dirSync({unsafeCleanup: true});
-
-            // create example theme
-            fs.mkdirSync(join(themePath.name, 'partials'));
-            fs.writeFileSync(join(themePath.name, 'index.hbs'));
-            fs.writeFileSync(join(themePath.name, 'partials', 'navigation.hbs'));
-
-            readDirectory(themePath.name)
-                .then(function (tree) {
-                    tree.should.eql({
-                        partials: {
-                            'navigation.hbs': join(themePath.name, 'partials', 'navigation.hbs')
-                        },
-                        'index.hbs': join(themePath.name, 'index.hbs')
-                    });
-
-                    done();
-                })
-                .catch(done)
-                .finally(themePath.removeCallback);
-        });
-
-        it('should read directory and ignore unneeded items', function (done) {
-            var themePath = tmp.dirSync({unsafeCleanup: true});
-
-            // create example theme
-            fs.mkdirSync(join(themePath.name, 'partials'));
-            fs.writeFileSync(join(themePath.name, 'index.hbs'));
-            fs.writeFileSync(join(themePath.name, 'partials', 'navigation.hbs'));
-
-            // create some trash
-            fs.mkdirSync(join(themePath.name, 'node_modules'));
-            fs.mkdirSync(join(themePath.name, 'bower_components'));
-            fs.mkdirSync(join(themePath.name, '.git'));
-            fs.writeFileSync(join(themePath.name, '.DS_Store'));
-
-            readDirectory(themePath.name, {ignore: ['.git']})
-                .then(function (tree) {
-                    tree.should.eql({
-                        partials: {
-                            'navigation.hbs': join(themePath.name, 'partials', 'navigation.hbs')
-                        },
-                        'index.hbs': join(themePath.name, 'index.hbs')
-                    });
-
-                    done();
-                })
-                .catch(done)
-                .finally(themePath.removeCallback);
-        });
-
-        it('should read directory and parse package.json files', function (done) {
-            var themePath, pkgJson;
-
-            themePath = tmp.dirSync({unsafeCleanup: true});
-            pkgJson = JSON.stringify({
-                name: 'test',
-                version: '0.0.0'
-            });
-
-            // create example theme
-            fs.mkdirSync(join(themePath.name, 'partials'));
-            fs.writeFileSync(join(themePath.name, 'package.json'), pkgJson);
-            fs.writeFileSync(join(themePath.name, 'index.hbs'));
-            fs.writeFileSync(join(themePath.name, 'partials', 'navigation.hbs'));
-
-            readDirectory(themePath.name)
-                .then(function (tree) {
-                    tree.should.eql({
-                        partials: {
-                            'navigation.hbs': join(themePath.name, 'partials', 'navigation.hbs')
-                        },
-                        'index.hbs': join(themePath.name, 'index.hbs'),
-                        'package.json': {
-                            name: 'test',
-                            version: '0.0.0'
-                        }
-                    });
-
-                    done();
-                })
-                .catch(done)
-                .finally(themePath.removeCallback);
-        });
-
-        it('should read directory and ignore invalid package.json files', function (done) {
-            var themePath, pkgJson;
-
-            themePath = tmp.dirSync({unsafeCleanup: true});
-            pkgJson = JSON.stringify({
-                name: 'test'
-            });
-
-            // create example theme
-            fs.mkdirSync(join(themePath.name, 'partials'));
-            fs.writeFileSync(join(themePath.name, 'package.json'), pkgJson);
-            fs.writeFileSync(join(themePath.name, 'index.hbs'));
-            fs.writeFileSync(join(themePath.name, 'partials', 'navigation.hbs'));
-
-            readDirectory(themePath.name)
-                .then(function (tree) {
-                    tree.should.eql({
-                        partials: {
-                            'navigation.hbs': join(themePath.name, 'partials', 'navigation.hbs')
-                        },
-                        'index.hbs': join(themePath.name, 'index.hbs'),
-                        'package.json': null
-                    });
-
-                    done();
-                })
-                .catch(done)
-                .finally(themePath.removeCallback);
-        });
-    });
-
-    describe('read-themes', function () {
-        it('should read directory and include only folders', function (done) {
-            var themesPath = tmp.dirSync({unsafeCleanup: true});
-
-            // create trash
-            fs.writeFileSync(join(themesPath.name, 'casper.zip'));
-            fs.writeFileSync(join(themesPath.name, '.DS_Store'));
-
-            // create actual theme
-            fs.mkdirSync(join(themesPath.name, 'casper'));
-            fs.mkdirSync(join(themesPath.name, 'casper', 'partials'));
-            fs.writeFileSync(join(themesPath.name, 'casper', 'index.hbs'));
-            fs.writeFileSync(join(themesPath.name, 'casper', 'partials', 'navigation.hbs'));
-
-            readThemes(themesPath.name)
-                .then(function (tree) {
-                    tree.should.eql({
-                        casper: {
-                            partials: {
-                                'navigation.hbs': join(themesPath.name, 'casper', 'partials', 'navigation.hbs')
-                            },
-                            'index.hbs': join(themesPath.name, 'casper', 'index.hbs')
-                        }
-                    });
-
-                    done();
-                })
-                .catch(done)
-                .finally(themesPath.removeCallback);
         });
     });
